# docs/swagger.yaml
basePath: /
definitions:
...
  traceability.ResTrackingValidation:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/traceability.Tracking'
        type: array
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
  traceability.Traceability:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  traceability.Tracking:
    properties:
      client_id:
        type: integer
      created_at:
        type: string
      expired_at:
        type: string
      id:
        type: integer
      max_num_validation:
        type: integer
      request_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  traceability.resTraceability:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/traceability.Traceability'
        type: array
      error:
        type: boolean
      msg:
        type: string
      type:
        type: integer
    type: object
...
host: 172.147.77.149:50050
info:
  contact:
    email: luis.lucero@bjungle.net
    name: API Support
  description: Api para OnBoarding y validación de identidad
  license:
    name: Software Owner
    url: https://www.bjungle.net/terms/licenses
  termsOfService: https://www.bjungle.net/terms/
  title: Check ID OnBoarding
  version: "1.4"
paths:
...
  /api/v1/traceability/user-session/{userID}:
    get:
      consumes:
      - application/json
      description: Método para obtener la trazabilidad registrada para el proceso
        de verificación de identidad
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del usuario
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/traceability.resTraceability'
      summary: Obtención de los datos de trazabilidad
      tags:
      - Traceability
  /api/v1/traceability/validation-identity/{id}:
    get:
      consumes:
      - application/json
      description: Método que obtiene el historial de validaciones de identidad de
        un usuario por su id
      parameters:
      - default: Bearer <Add access token here>
        description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/traceability.ResTrackingValidation'
      summary: Obtiene el historial de validaciones de identidad de un usuario
      tags:
      - Traceability
...
swagger: "2.0"
...